Generalizo: Mapa y Disposicion heredan de nueva clase ConjuntoDeCasilleros.
